requirements.txt
pywhatkit
mysql-connector-python